Split dpo indicator into several files.
