+++ js/masters/indicators/indicators-all.src.js
@@ -21,7 +21,7 @@ import '../../Stock/Indicators/ATRIndicator.js';
 import '../../Stock/Indicators/BBIndicator.js';
 import '../../Stock/Indicators/CCIIndicator.js';
 import '../../Stock/Indicators/CMF/CMFIndicator.js';
-import '../../Stock/Indicators/DPOIndicator.js';
+import '../../Stock/Indicators/DPO/DPOIndicator.js';
 import '../../Stock/Indicators/EMAIndicator.js';
 import '../../Stock/Indicators/ChaikinIndicator.js';
 import '../../Stock/Indicators/DEMAIndicator.js';
