+++ ts/Stock/Indicators/DPO/DPOOptions.d.ts
@@ -1,0 +1,40 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type {
+    SMAOptions,
+    SMAParamsOptions
+} from '../SMA/SMAOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface DPOOptions extends SMAOptions {
+    params?: DPOParamsOptions;
+}
+
+export interface DPOParamsOptions extends SMAParamsOptions {
+    // for inheritance
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default DPOOptions;
