+++ ts/Stock/Indicators/DPO/DPOPoint.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import DPOIndicator from './DPOIndicator';
+import type SMAPoint from '../SMA/SMAPoint';
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+declare class DPOPoint extends SMAPoint {
+    public series: DPOIndicator;
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default DPOPoint;
